@lemire 2020-07-16
Comments are invited.