@lemire 2020-07-17
Your reasoning is sound, but I'd really want other folks to discuss this because it is not going to be a net positive.