@lemire 2020-07-24
> The main benefit is that if you OR together a bunch of containers that each stay below 4096 you'll only do the expensive thing once. The current implementation will repeatedly map to bitmap containers and then convert back. It's a lot harder to imagine a reasonable antagonistic case where you pay a penalty for not preemptively converting to a bitmap.

That sounds credible.

Your argument is that we should be optimistic that we won't need to do new allocation. Is that it?