@jacksonrnewhouse 2020-07-24
Yep, that's the idea.