@jacksonrnewhouse 2020-10-21
Took another stab at this I'll be putting up in a new PR.